+++ public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link rel="icon" href="%PUBLIC_URL%/insta_favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Westagram</title>
   </head>
