+++ src/pages/ohj/Main/components/Article.scss
@@ -1,0 +1,275 @@
+$whiteColor: #fff;
+
+@mixin displayFlex($justify) {
+  display: flex;
+  justify-content: $justify;
+}
+
+.article {
+  width: 64%;
+
+  .feeds {
+    width: 500px;
+    height: auto;
+    margin-top: 57px;
+    background-color: $whiteColor;
+    border: 1px solid #ddd;
+    border-radius: 3px;
+
+    /* post_host 영역 */
+
+    #post_host {
+      height: 55px;
+      background-color: $whiteColor;
+
+      position: relative;
+
+      .post_host_pic {
+        display: inline-block;
+        width: 31px;
+        height: 31px;
+        margin: 12px 0 0 15px;
+      }
+
+      .post_host_pic img {
+        width: 100%;
+        border-radius: 50%;
+      }
+
+      .post_host_nickname {
+        display: inline-block;
+      }
+
+      .post_host_nickname span {
+        font-family: '나눔고딕';
+        font-size: 13px;
+        font-weight: 600;
+        color: #000;
+
+        position: absolute;
+        left: 60px;
+        top: 21px;
+      }
+
+      .dots {
+        display: inline-block;
+        float: right;
+        overflow: hidden;
+        margin: 16px 20px 0 0;
+      }
+
+      .dot_black {
+        display: inline-block;
+        width: 3px;
+        height: 3px;
+        margin-left: 5px;
+        background-color: #222;
+        border-radius: 50%;
+      }
+    }
+
+    /* post_pic 영역 */
+
+    #post_pic {
+      width: inherit;
+      height: 400px;
+      overflow: hidden;
+      background-color: blue;
+
+      img {
+        width: 100%;
+        translate: 0 -87px;
+      }
+    }
+
+    /* post_txt 영역 */
+
+    .post_txt_icons {
+      @include displayFlex(space-between);
+
+      width: 100%;
+      height: 100%;
+      padding-top: 11px;
+
+      .post_txt_icon_left span {
+        display: inline-block;
+        width: 22px;
+        height: 22px;
+      }
+
+      .post_txt_icon1 img {
+        width: 100%;
+        margin: 1px 2px 0 15px;
+        cursor: pointer;
+      }
+
+      .post_txt_icon2 {
+        margin-left: 24px;
+      }
+
+      .post_txt_icon2 img {
+        width: 100%;
+      }
+
+      .post_txt_icon3 {
+        margin-left: 10px;
+      }
+
+      .post_txt_icon3 img {
+        width: 100%;
+      }
+
+      .post_txt_icon_right span {
+        display: inline-block;
+        width: 23px;
+        height: 23px;
+      }
+
+      .post_txt_icon4 {
+        margin: 1px 12px 0 0;
+      }
+
+      .post_txt_icon4 img {
+        width: 100%;
+      }
+    }
+
+    .post_txt_comment {
+      position: relative;
+      background-color: $whiteColor;
+      padding-left: 8px;
+      padding-bottom: 7px;
+
+      a {
+        display: inline-block;
+        width: 25px;
+        height: 25px;
+        margin: 10px;
+      }
+
+      img {
+        width: 100%;
+        border-radius: 50%;
+      }
+
+      .like {
+        font-family: '나눔고딕';
+        font-size: 13px;
+        position: absolute;
+        left: 50px;
+        top: 16px;
+      }
+
+      .like .strong {
+        font-weight: 600;
+      }
+
+      .comment {
+        position: relative;
+
+        li {
+          font-family: '나눔고딕';
+          font-size: 14px;
+          margin-left: 17px;
+        }
+
+        .post_host_nickname {
+          width: 300px;
+          text-overflow: ellipsis;
+          white-space: nowrap;
+          overflow: hidden;
+        }
+
+        .strong {
+          font-weight: 600;
+        }
+
+        .more {
+          font-size: 14px;
+          color: #777;
+
+          position: absolute;
+          left: 325px;
+          top: 2px;
+        }
+
+        .commentor_nickname {
+          @include displayFlex(space-between);
+        }
+
+        .comment1 {
+          margin: 10px 0 7px 17px;
+          font-size: 13px;
+        }
+
+        .comment2 {
+          margin-left: 17px;
+          padding-bottom: 7px;
+          font-size: 13px;
+        }
+
+        .comment3 {
+          margin-left: 17px;
+          padding-bottom: 7px;
+          font-size: 13px;
+        }
+
+        img {
+          width: 10px;
+          height: 10px;
+          margin-right: 20px;
+          margin-top: 2px;
+        }
+      }
+    }
+
+    .comment_window {
+      @include displayFlex(space-between);
+
+      background-color: $whiteColor;
+
+      border-top: 1px solid #eee;
+
+      input.comment_write {
+        width: 100%;
+        padding: 16px 0 15px 25px;
+        border: none;
+      }
+
+      input.comment_write::placeholder {
+        font-family: '나눔고딕';
+        color: #bbb;
+      }
+
+      input.comment_write:focus {
+        outline: none;
+      }
+
+      .activePublish {
+        float: right;
+        width: 50px;
+        background-color: $whiteColor;
+        border: none;
+        margin-right: 15px;
+
+        font-family: '나눔고딕';
+        font-weight: 600;
+        color: royalblue;
+        cursor: pointer;
+      }
+
+      .noActivePublish {
+        float: right;
+        width: 50px;
+        background-color: $whiteColor;
+        border: none;
+        margin-right: 15px;
+
+        font-family: '나눔고딕';
+        font-weight: 600;
+        color: royalblue;
+        opacity: 0.3;
+      }
+    }
+  }
+}
