+++ src/pages/ohj/Main/components/CommentList.js
@@ -1,17 +1,15 @@
 import React, { useState } from 'react';
 import OhjLogin from '../../Login/OhjLogin';
 
-function CommentList(props) {
+function CommentList({ userName, userComment }) {
   return (
-    <div>
-      <li className="commentor_nickname comment3">
-        <span>
-          <span className="strong">{props.userName}</span>
-          <span>{props.userComment}</span>
-        </span>
-        <img src="images/ohj/like_heart.png" alt="댓글 좋아요" />
-      </li>
-    </div>
+    <li className="commentor_nickname comment3">
+      <span>
+        <span className="strong">{userName}</span>
+        <span>{userComment}</span>
+      </span>
+      <img src="images/ohj/like_heart.png" alt="댓글 좋아요" />
+    </li>
   );
 }
